requirements.txt
aws-cdk-lib==2.32.0
constructs>=10.0.0,<11.0.0

pyspark~=3.3.0
boto3~=1.24.86
Faker~=15.0.0
requests~=2.28.1
botocore~=1.27.86
pandas~=1.5.0
pyarrow~=9.0.0
psutil~=5.9.2
cloudwatch>=1.0.5
aws-cdk.aws-apigatewayv2-integrations-alpha 